adjustments to lexor and ST lab 3

+++ Java/rescueClean/src/solution/CurrentLocation.java
@@ -2,10 +2,10 @@ package solution;
 
 public class CurrentLocation {
 
-    int Building;
+    Long Building;
     double totalDistance;
 
-    CurrentLocation(int Building, double totalDistance) {
+    CurrentLocation(Long Building, double totalDistance) {
         this.Building = Building;
         this.totalDistance = totalDistance;
     }
